+++ packages/server/lib/cloud/api/index.ts
@@ -458,17 +458,19 @@ export default {
       'platform',
     ])
 
-    return rp.post({
-      body,
-      url: recordRoutes.instances(runId),
-      json: true,
-      encrypt: preflightResult.encrypt,
-      timeout: timeout ?? SIXTY_SECONDS,
-      headers: {
-        'x-route-version': '5',
-        'x-cypress-run-id': runId,
-        'x-cypress-request-attempt': 0,
-      },
+    return retryWithBackoff((attemptIndex) => {
+      return rp.post({
+        body,
+        url: recordRoutes.instances(runId),
+        json: true,
+        encrypt: preflightResult.encrypt,
+        timeout: timeout ?? SIXTY_SECONDS,
+        headers: {
+          'x-route-version': '5',
+          'x-cypress-run-id': runId,
+          'x-cypress-request-attempt': attemptIndex,
+        },
+      })
     })
     .catch(RequestErrors.StatusCodeError, formatResponseBody)
     .catch(tagError)